@@ -8,13 +8,13 @@ buildscript {
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
+    compileSdkVersion 30
+    buildToolsVersion "30.0.2"
 
     defaultConfig {
         applicationId "com.lloydtorres.stately"
         minSdkVersion 16
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 45
         versionName "1.10.3"
         vectorDrawables.useSupportLibrary = true
@@ -39,20 +39,20 @@ dependencies {
         exclude module: 'stax-api'
         exclude module: 'xpp3'
     }
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:cardview-v7:28.0.0'
-    implementation 'com.android.support:preference-v7:28.0.0'
-    implementation 'com.android.support:preference-v14:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.preference:preference:1.1.1'
+    implementation 'androidx.legacy:legacy-preference-v14:1.0.0'
     implementation 'com.android.volley:volley:1.1.1'
-    implementation 'com.squareup.picasso:picasso:2.71828'
     implementation 'com.github.satyan:sugar:1.5'
     implementation 'com.github.omadahealth:swipy:1.2.3@aar'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
+    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'com.jpardogo.materialtabstrip:library:1.1.1'
+    implementation 'com.r0adkll:slidableactivity:2.0.6'
+    implementation 'com.squareup.picasso:picasso:2.71828'
     implementation 'org.atteo:evo-inflector:1.2.2'
     implementation 'org.jsoup:jsoup:1.13.1'
     implementation 'org.sufficientlysecure:html-textview:3.9'
-    implementation 'com.jpardogo.materialtabstrip:library:1.1.1'
-    implementation 'com.r0adkll:slidableactivity:2.0.6'
     implementation 'pl.bclogic:pulsator4droid:1.0.3'
 }
