@@ -17,7 +17,9 @@
 package com.lloydtorres.stately.core;
 
 import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -29,7 +31,7 @@ import com.lloydtorres.stately.helpers.RaraHelper;
  * Created by Lloyd on 2016-09-12.
  * Skeleton for sub-fragments that use the recycler layout.
  */
-public abstract class RecyclerSubFragment extends DetachFragment {
+public abstract class RecyclerSubFragment extends Fragment {
 
     protected RecyclerView mRecyclerView;
     protected RecyclerView.LayoutManager mLayoutManager;
