@@ -17,9 +17,11 @@
 package com.lloydtorres.stately.core;
 
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+
+import androidx.fragment.app.Fragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -31,7 +33,7 @@ import com.lloydtorres.stately.helpers.RaraHelper;
  * Created by Lloyd on 2016-09-11.
  * Skeleton for fragments that use the Refreshview layout.
  */
-public abstract class RefreshviewFragment extends DetachFragment {
+public abstract class RefreshviewFragment extends Fragment {
 
     protected View mView;
     protected Toolbar toolbar;
