@@ -48,8 +48,8 @@ package com.lloydtorres.stately.helpers;
 
 import android.content.Context;
 import android.content.res.Configuration;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 import android.view.View;
 
 import com.github.mikephil.charting.charts.LineChart;
