@@ -53,8 +53,8 @@ import android.content.Intent;
 import android.graphics.Typeface;
 import android.net.Uri;
 import android.os.Build;
-import android.support.design.widget.Snackbar;
-import android.support.v4.app.FragmentManager;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.fragment.app.FragmentManager;
 import android.text.Html;
 import android.text.Spannable;
 import android.text.SpannableString;
