@@ -16,7 +16,7 @@
 
 package com.lloydtorres.stately.helpers;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
 
