@@ -19,8 +19,8 @@ package com.lloydtorres.stately.helpers.dialogs;
 import android.app.Dialog;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AppCompatDialog;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AppCompatDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
