@@ -19,8 +19,8 @@ package com.lloydtorres.stately.issues;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v7.app.AlertDialog;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+import androidx.appcompat.app.AlertDialog;
 
 import com.android.volley.NetworkError;
 import com.android.volley.NoConnectionError;
