@@ -19,9 +19,9 @@ package com.lloydtorres.stately.region;
 import android.content.Context;
 import android.graphics.Typeface;
 import android.os.Parcelable;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.FragmentManager;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.text.Spannable;
 import android.text.SpannableString;
 import android.text.Spanned;
