@@ -18,8 +18,8 @@ package com.lloydtorres.stately.region;
 
 import android.content.Context;
 import android.os.Parcelable;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.FragmentManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
