@@ -18,11 +18,13 @@ package com.lloydtorres.stately.telegrams;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.v4.app.FragmentManager;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.appcompat.app.AlertDialog;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -37,7 +39,6 @@ import com.android.volley.Response;
 import com.android.volley.TimeoutError;
 import com.android.volley.VolleyError;
 import com.lloydtorres.stately.R;
-import com.lloydtorres.stately.core.DetachFragment;
 import com.lloydtorres.stately.core.IToolbarActivity;
 import com.lloydtorres.stately.core.StatelyActivity;
 import com.lloydtorres.stately.dto.Telegram;
@@ -66,7 +67,7 @@ import java.util.regex.Matcher;
  * Created by Lloyd on 2016-03-08.
  * This is the Telegrams section of the main Stately activity.
  */
-public class TelegramsFragment extends DetachFragment {
+public class TelegramsFragment extends Fragment {
     public static final String KEY_PAST_OFFSET = "keyPastOffset";
     public static final String KEY_TELEGRAMS = "keyTelegrams";
     public static final String KEY_FOLDERS = "keyFolders";
