@@ -17,7 +17,7 @@
 package com.lloydtorres.stately.zombie;
 
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.android.volley.NetworkError;
 import com.android.volley.NoConnectionError;
